pyproject.toml
[project]
name = "vggsfm"
version = "2.0.0"
authors = [{ name = "Jianyuan Wang" }]
description = "A package for the VGGSfM project."
readme = { file = "README.md", content-type = "text/markdown" }
license = { file = "LICENSE.txt" }
urls = { homepage = "https://github.com/facebookresearch/vggsfm.git" }
requires-python = ">=3.10"
dependencies = [
    "numpy (>=1.26.0,<2.0.0)",
    "scipy (>=1.15.3,<2.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "einops (>=0.8.1,<0.9.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "plotly (>=6.1.2,<7.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "imageio[ffmpeg] (>=2.37.0,<3.0.0)",
    "gradio (>=5.34.2,<6.0.0)",
    "trimesh (>=4.6.12,<5.0.0)",
    "huggingface-hub (>=0.33.0,<0.34.0)",
    "pycolmap (>=3.11.1,<4.0.0)",
    "pyceres (>=2.4,<3.0)",
    "poselib (>=2.0.4,<3.0.0)",
    "fvcore (>=0.1.5.post20221221,<0.2.0)",
    "iopath (>=0.1.10,<0.2.0)",
    "joblib (>=1.5.1,<2.0.0)",
    "coloredlogs (>=15.0.1,<16.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.cpu]
optional = true

[tool.poetry.group.cpu.dependencies]
torch = [
    { version = "2.7.0+cpu", markers = "extra != 'cuda'", source = "pytorch-cpu" },
]
torchvision = [
    { version = "0.22.0+cpu", markers = "extra != 'cuda'", source = "pytorch-cpu" },
]

[tool.poetry.group.cuda]
optional = true

[tool.poetry.group.cuda.dependencies]
torch = [
    { version = "2.7.0+cu126", markers = "extra == 'cuda'", source = "pytorch-cuda" },
]
torchvision = [
    { version = "0.22.0+cu126", markers = "extra == 'cuda'", source = "pytorch-cuda" },
]   
xformers = [
    { version = "0.0.30", markers = "extra == 'cuda'", source = "pytorch-cuda" },
]

[tool.poetry.dependencies]
lightglue = { path = "dependency/LightGlue", develop = true }

[tool.poetry.group.dev.dependencies]
flake8-pyproject = "^1.2.3"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.16.1"


[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[[tool.poetry.source]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu126"
priority = "explicit"


[build-system]
requires = ["setuptools", "poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
length_sort = true
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = false
use_parentheses = true
known_third_party = ["numpy", "torch", "wandb"]
skip_glob = [
    "**/build/**",
    "**/.mypy_cache/**",
    "**/.pytest_cache/**",
    "**/.vscode/**",
]

[tool.mypy]
python_version = "3.11"
follow_imports = "silent"
plugins = ["numpy.typing.mypy_plugin"]
disallow_any_generics = true
strict_optional = true
check_untyped_defs = true
allow_redefinition = true

[tool.flake8]
max-line-length = 120
ignore = "Q000,E127,E203,E402,W503,W504"

[tool.black]
target-version = ["py311"]
line-length = 120
skip-string-normalization = true
